requirements.txt
# Core SDK
aixplain>=0.3.0

# Env management
python-dotenv>=1.0.1

# Utils
requests>=2.32.0
pypdf>=5.0.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
pandas>=2.2.2